+++ docs/source/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 * Mako is now installed by default, and the ``mako`` extra is gone. If you installed Liara using ``liara[mako]``, please switch to ``liara`` going forward.
 * :py:attr:`liara.template.Page.content` now returns an empty string for :py:class:`~liara.nodes.IndexNode` instances. Previously, it would raise an exception.
 * ``liara quickstart`` gained a new option, ``--template-backend``, which allows selecting between ``jinja2`` and ``mako`` templates.
+* The collection sort order can be reversed now. See :doc:`collections` for more details.
 
 2.2.0
 -----